removed Chart as we do not override anything
