+++ src/components/index.ts
@@ -12,11 +12,8 @@ export * from './buttons/BackToButton'
 export { default as UploadButton } from './buttons/UploadButton'
 export * from './buttons/UploadButton'
 
-export { default as StatsChart } from './charts/StatChart'
-export * from './charts/StatChart'
-
-export { default as Chart } from './charts/Chart'
-export * from './charts/Chart'
+export { default as StatsChart } from './statChart'
+export * from './statChart'
 
 export { default as Typography } from './dataDisplay/Typography'
 export * from './dataDisplay/Typography'
