+++ src/components/statChart/index.ts
