+++ src/stories/statChart/LinePreview.tsx
@@ -2,7 +2,7 @@
 // This source code is licensed under the MIT license.
 
 import React from 'react'
-import { StatsChart } from 'components'
+import { Button, StatsChart } from 'components'
 import AccessTime from '@mui/icons-material/AccessTime'
 import Grid from '@mui/material/Grid2'
 
@@ -22,23 +22,29 @@ const LinePreview = () => {
             type: 'line',
             xAxis: [{ data: [1, 2, 3, 5, 8, 10] }],
             series: [{ data: [2, 5.5, 2, 8.5, 1.5, 5] }],
-            width: 500,
+            width: 450,
             height: 300
           }}
         />
       </Grid>
-      {/* <Grid size={6}>
+      <Grid size={6}>
         <StatsChart
-          type="line"
-          data={statsChartData}
-          chartColor={'info'}
           StatIcon={AccessTime}
           statText={'Line chart infos'}
-          title={'Line Chart with action button'}
-          options={statsChartOptions}
+          title={'Line Chart'}
+          chartColor="secondary"
+          iconColor="info"
+          text={'This is a line chart with action button'}
           statAction={<Button size="tiny">Ok</Button>}
+          chart={{
+            type: 'line',
+            xAxis: [{ data: [1, 2, 3, 5, 8, 10] }],
+            series: [{ data: [2, 5.5, 2, 8.5, 1.5, 5] }],
+            width: 450,
+            height: 300
+          }}
         />
-      </Grid> */}
+      </Grid>
     </Grid>
   )
 }
