+++ src/stories/statChart/ScatterPreview.tsx
@@ -5,31 +5,33 @@ import React from 'react'
 import { StatsChart } from 'components'
 import AccessTime from '@mui/icons-material/AccessTime'
 import Grid from '@mui/material/Grid2'
+import { scatterData } from './_mocks'
 
-const PiePreview = () => {
+const ScatterPreview = () => {
   return (
     <Grid container spacing={2}>
       <Grid size={6}>
         <StatsChart
           StatIcon={AccessTime}
-          statText={'Pie chart infos'}
-          title={'Pie Chart'}
+          statText={'Scatter chart infos'}
+          title={'Scatter Chart'}
           chartColor="info"
           iconColor="info"
-          text={'This is a pie chart'}
+          text={'This is a scatter chart'}
           statAction={'View'}
           chart={{
-            type: 'pie',
+            type: 'scatter',
             series: [
               {
-                data: [
-                  { id: 0, value: 10, label: 'series A' },
-                  { id: 1, value: 15, label: 'series B' },
-                  { id: 2, value: 20, label: 'series C' }
-                ]
+                label: 'Series A',
+                data: scatterData.map(v => ({ x: v.x1, y: v.y1, id: v.id }))
+              },
+              {
+                label: 'Series B',
+                data: scatterData.map(v => ({ x: v.x1, y: v.y2, id: v.id }))
               }
             ],
-            width: 500,
+            width:450,
             height: 300
           }}
         />
@@ -38,4 +40,4 @@ const PiePreview = () => {
   )
 }
 
-export default PiePreview
+export default ScatterPreview
