+++ src/stories/statChart/TypesPreview.tsx
